build: zero-warnings policy + replace placeholder stubs with real impls
Two interlocked changes that close out skeleton-phase shortcuts:

REAL IMPLEMENTATIONS (no more placeholder stubs printing [stub] at runtime)

- AxisToQuat: full Shepperd's-method matrix-to-quaternion conversion,
  numerically stable variant. Replaces identity-quaternion placeholder
  that was producing wrong bone/anim/ragdoll rotations.
- Sys_GetValue / Sys_SetValue: real pthread_key_create-backed TLS, 16
  slots, lazy init via std::call_once. Replaces null-returning stub.
- va: 32 rotating buffer slots (was 8) — matches upstream so call chains
  with up to 32 outstanding strings work safely.

ZERO-WARNINGS POLICY ENFORCED

- Removed -Wno-sign-compare suppression from both POSIX and Switch builds.
- Added -Werror to both. Build fails on any warning from now on.
- All previously-suppressed warnings fixed at the site with explicit casts:
  - base64.cpp: c declared int (was unsigned int) so EOF compare works;
    encode loop counter changed to unsigned to match in_size.
  - aabbtree.cpp: 0x80000000 cast to int for INT_MIN sentinel comparison;
    options->itemCount cast to unsigned for 0x40u compare.
  - com_constantconfigstrings.cpp: configStringHash cast to unsigned for
    hash equality; index compared against (unsigned)count.
  - com_convexhull.cpp: pointCount != (unsigned)(botIndex + 1) cast.
  - q_parse.cpp: startDepth == (unsigned)iMaxNesting cast, plus
    parentheses on 9 '&&'-within-'||' precedence warnings in the parser.
  - com_shared.cpp: Com_Prefetch wrapped in same #if 0 as its only call
    site (the disabled hex-rays optimized Com_Memcpy body); dest/src
    locals cast to void to silence -Wunused-variable in Release builds
    where iassert() expands to nothing.
- Headers with intentionally-unused static const char* name tables
  marked [[maybe_unused]]: bg_public.h (entityTypeNames, eventnames),
  qcommon.h (WeaponStateNames).

Project memory updated to enforce this policy on future commits — see
~/.claude/projects/.../feedback_zero_warnings.md.

docs/RISKS.md updated: entries 3 (Sys_GetValue), 4 (AxisToQuat), and 5
(sign-compare suppression) all marked RESOLVED.

+++ src/universal/aabbtree.cpp
@@ -176,7 +176,7 @@ static bool PickAabbSplitPlane(
             }
         }
     }
-    return bestHeuristic != 0x80000000;
+    return bestHeuristic != (int)0x80000000;
 }
 
 double __cdecl AddedVolume(const float *addedmins, const float *addedmaxs, const float *mins, const float *maxs)
@@ -400,7 +400,7 @@ int __cdecl BuildAabbTree(const GenericAabbTreeOptions *options)
     float sortedBounds[3][64]; // [esp+150h] [ebp-308h] BYREF
     unsigned __int8 *itemCopies; // [esp+454h] [ebp-4h]
 
-    if (options->itemCount > 0x40u)
+    if ((unsigned int)options->itemCount > 0x40u)
     {
         remap = (int*)operator new(4 * options->itemCount);
         sortedMins = (float*)operator new(4 * options->itemCount);
